FST: Avoid compiler warning

+++ fst/ScanDir.cc
@@ -170,7 +170,7 @@ ScanDir::SetConfig(const std::string& key, long long value)
   } else if (key == eos::common::SCAN_ENTRY_INTERVAL_NAME) {
     mEntryIntervalSec = value;
   } else if (key == eos::common::SCAN_DISK_INTERVAL_NAME) {
-    if (mDiskIntervalSec != value) {
+    if (mDiskIntervalSec != static_cast<uint64_t>(value)) {
       mDiskIntervalSec = value;
       mDiskThread.join();
       mDiskThread.reset(&ScanDir::RunDiskScan, this);
@@ -178,7 +178,7 @@ ScanDir::SetConfig(const std::string& key, long long value)
   } else if (key == eos::common::SCAN_NS_INTERVAL_NAME) {
 #ifndef _NOOFS
 
-    if (mNsIntervalSec != value) {
+    if (mNsIntervalSec != static_cast<uint64_t>(value)) {
       mNsIntervalSec = value;
       mNsThread.join();
       mNsThread.reset(&ScanDir::RunNsScan, this);
